# contracts/api.yaml
...
openapi: 3.0.3
info:
  title: "AML Case Management API"
  version: "1.0.0"
  description: |
    API contracts for the AML Case Management System.

    ## Overview

    This API enables Spriggy to meet AUSTRAC AML/CTF compliance obligations by providing
    endpoints for managing KYC remediation, PEP/Sanctions screening disposition, and
    suspicious activity investigation workflows.

    ## Authentication

    All endpoints require Bearer token authentication via OIDC SSO integration (D5).
    Include the token in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    Webhook endpoints use HMAC signature validation instead of Bearer tokens (D11).

    ## Roles

    The system enforces four distinct roles (FR-063):
    - **L1_ANALYST**: Triage and basic case handling
    - **L2_ANALYST**: Full investigation and SMR recommendation authority
    - **AML_MANAGER**: Approval authority for SMRs
    - **READ_ONLY**: View and export only

    ## Error Handling

    All errors follow a consistent format with machine-readable codes.
    See the ErrorResponse schema for details.

    ## Optimistic Locking

    Case mutations require a `version` field for optimistic locking (D8).
    Version mismatch returns 409 Conflict.

    ## Constitution Alignment

    - **Principle I**: All mutating endpoints create audit log entries
    - **Principle II**: Endpoints enforce role-based access
...
  contact:
    name: AML Compliance Team
...
servers:
  - url: /api/v1
...
paths:
...
  /webhooks/greenid:
    post:
      summary: Receive GreenID screening alert
      description: |
        Webhook endpoint for GreenID PEP/Sanctions screening alerts.

        **Authentication**: HMAC signature in X-Signature-256 header (D11)

        **Duplicate Handling**: Webhooks within 24-hour window for same
        customer/alert-type are linked to existing case (EC-005, D3).

        **Source Requirements**: FR-001, FR-002, FR-029, FR-030, FR-031, US-6, US-7
      operationId: receiveGreenIdWebhook
      tags:
        - Webhooks
      security: []
      parameters:
        - $ref: '#/components/parameters/WebhookSignature'
        - $ref: '#/components/parameters/WebhookTimestamp'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GreenIdWebhookPayload'
      responses:
        '202':
          description: Webhook accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: WEBHOOK_AUTH_FAILED
                message: Invalid or missing webhook signature
        '409':
          description: Duplicate webhook (already processed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookDuplicateResponse'
  /webhooks/indue:
    post:
      summary: Receive Indue transaction monitoring alert
      description: |
        Webhook endpoint for Indue transaction monitoring alerts.

        **Authentication**: HMAC signature in X-Signature-256 header (D11)

        **Source Requirements**: FR-001, FR-002, US-3
      operationId: receiveIndueWebhook
      tags:
        - Webhooks
      security: []
      parameters:
        - $ref: '#/components/parameters/WebhookSignature'
        - $ref: '#/components/parameters/WebhookTimestamp'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndueWebhookPayload'
      responses:
        '202':
          description: Webhook accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: WEBHOOK_AUTH_FAILED
                message: Invalid or missing webhook signature
        '409':
          description: Duplicate webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookDuplicateResponse'

  # --------------------------------------------------------------------------
  # Case Endpoints
  # --------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from OIDC SSO

  # --------------------------------------------------------------------------
  # Common Parameters
  # --------------------------------------------------------------------------
  parameters:
...
    WebhookSignature:
      name: X-Signature-256
      in: header
      required: true
      description: HMAC-SHA256 signature of request body
      schema:
        type: string
    WebhookTimestamp:
      name: X-Timestamp
      in: header
      required: true
      description: ISO8601 timestamp of webhook generation
      schema:
        type: string
        format: date-time

  # --------------------------------------------------------------------------
  # Common Responses
  # --------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable message
          example: One or more fields are invalid
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    GreenIdWebhookPayload:
      type: object
      required:
        - alertId
        - customerId
        - alertType
        - matchData
        - timestamp
      properties:
        alertId:
          type: string
          description: GreenID alert identifier
        customerId:
          type: string
          description: Spriggy customer identifier
        alertType:
          type: string
          enum: [SANCTIONS, PEP]
          description: Type of screening alert
        confidenceScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Match confidence score (PEP alerts only)
        matchData:
          type: object
          description: Screening match details
          properties:
            matchedName:
              type: string
            matchedList:
              type: string
            matchReasons:
              type: array
              items:
                type: string
        customerOnboardingStatus:
          type: string
          enum: [IN_PROGRESS, COMPLETED]
          description: Customer's onboarding status at alert time
        timestamp:
          type: string
          format: date-time
          description: Alert generation time
    IndueWebhookPayload:
      type: object
      required:
        - alertId
        - customerId
        - alertType
        - alertDetails
        - timestamp
      properties:
        alertId:
          type: string
          description: Indue alert identifier
        customerId:
          type: string
          description: Spriggy customer identifier
        alertType:
          type: string
          enum: [SUSPICIOUS_TRANSACTION, UNUSUAL_PATTERN, THRESHOLD_BREACH]
        alertDetails:
          type: object
          description: Transaction monitoring alert details
          properties:
            ruleTriggered:
              type: string
            transactionIds:
              type: array
              items:
                type: string
            totalAmount:
              type: number
            currency:
              type: string
        customerOnboardingStatus:
          type: string
          enum: [IN_PROGRESS, COMPLETED]
          description: |
            Customer's onboarding status when alert was triggered.
            - IN_PROGRESS: Customer is still being onboarded (onboarding alerts per FR-036)
            - COMPLETED: Customer onboarding is complete (existing customer alerts per FR-037)
            Used to distinguish between onboarding alerts (which may trigger blocks) and
            existing customer alerts (which require analyst discretion on restrictions).
        timestamp:
          type: string
          format: date-time
    WebhookAcceptedResponse:
      type: object
      required:
        - received
        - caseId
        - caseReference
      properties:
        received:
          type: boolean
          example: true
        caseId:
          type: string
          format: uuid
          description: Created case ID
        caseReference:
          type: string
          description: Human-readable case reference
          example: AML-1001
    WebhookDuplicateResponse:
      type: object
      required:
        - duplicate
        - existingCaseId
        - existingCaseReference
      properties:
        duplicate:
          type: boolean
          example: true
        existingCaseId:
          type: string
          format: uuid
        existingCaseReference:
          type: string
          example: AML-1001
        message:
          type: string
          example: Duplicate webhook linked to existing case

    # --- Case Schemas ---
...
security:
  - bearerAuth: []